AccountChecker.cpp
This file is to take in an account number then to ask for your balance , total items charged to customer, total of all credits by customer and allowed credit for that month. It would then check if the allowed credits is less than balance , if so then it gives an error message for credit limit exceeded, if not then the new balance is calculated by Balance + items minus the total credits.