DiceDriver.cpp
This file asks the user to enter the number of sides to their two dice separately and the number of times to roll. This creates an array of the length of the total number of dice rolls. It uses a random generator to roll the dice and get the outcomes. It then outputs the number of times a specific number has outputted.